Buscar termos nos campos textuais dos dados do Catálogo de Teses e Dissertações da CAPES

Esta função permite buscar termos específicos nos campos textuais de um `data.frame` previamente carregado.

buscar_texto_capes(dados, termo, campo)

Arguments

dados

`data.frame` contendo os dados do Catálogo de Teses e Dissertações da CAPES.

termo

string, termo a ser buscado.

campo

string, nome do campo onde buscar (ex.: "resumo", "titulo").

Value

um `data.frame` com as linhas que correspondem à busca ou uma mensagem informando que nenhum resultado foi encontrado.