Identifiers (IDs) on OSF for the annual data of the Catalog of Theses and Dissertations from the Brazilian Coordination for the Improvement of Higher Education Personnel (CAPES)

A data frame containing the years and the corresponding IDs for downloading the files.

Usage

years_osf

Format

A data frame with the following columns:

year

Year of the data (1987-2022).

osf_id

OSF ID corresponding to the year.

Source

https://osf.io/

Examples

data(years_osf)
head(years_osf)
#>   year osf_id
#> 1 1987  nsdf7
#> 2 1988  uc6vh
#> 3 1989  rxmg2
#> 4 1990  r5gs9
#> 5 1991  s4cny
#> 6 1992  cmzgv