Download CAPES Data

Downloads CAPES theses and dissertations data files from OSF for selected years.

Usage

download_capes_data(years, destination = tempdir(), timeout = 120)

baixar_dados_capes(years, destination = tempdir(), timeout = 120)

Arguments

years

A vector with the desired years.

destination

The directory where the files will be saved (default: temporary directory).

timeout

The timeout in seconds for the download process (default: 120 seconds).

Value

A list of file paths for the downloaded or already existing files.

Examples

# \donttest{
# Download data for the years 1987 and 1990
capes_files <- download_capes_data(c(1987, 1990))
#> Downloading: C:\Users\HUGO~1.VAS\AppData\Local\Temp\RtmpeqmqwL/capes_1987.parquet
#> Downloading: C:\Users\HUGO~1.VAS\AppData\Local\Temp\RtmpeqmqwL/capes_1990.parquet
# }